requirements.txt
Babel==2.12.1
click==8.1.3
Flask==2.2.3
flask-babel==3.0.1
Flask-Table==0.5.0
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
protobuf==3.20.3
PyMySQL==1.0.3
pytz==2022.7.1
Werkzeug==2.2.3
zipp==3.15.0
python-dotenv==1.0.0